Add-on searching, about:config support

+++ app/src/main/java/com/cookiejarapps/android/smartcookieweb/BaseBrowserFragment.kt
@@ -103,10 +103,10 @@ abstract class BaseBrowserFragment : Fragment(), UserInteractionHandler, Activit
             securityIconSecure = 0xFF5cb85c.toInt(),
             text = 0xFF0c0c0d.toInt(),
             menu = 0xFF20123a.toInt(),
-            separator = 0x1E15141a.toInt(),
+            separator = 0x1E15141a,
             trackingProtection = 0xFF20123a.toInt(),
             emptyIcon = 0xFF20123a.toInt(),
-            hint = 0x1E15141a.toInt()
+            hint = 0x1E15141a
         )
 
         layout.toolbar.elevation = 8f
